Make sure not to use LTO during build
When building a dependent crate with CXXFLAGS="-flto" we see a bunch of
undefined references to functions coming from libnitrokey [0]. E.g.,

>   = note: ld: nitrokey-rs/target/debug/deps/libnitrokey-f907de21b60ae79e.rlib(nitrokey-f907de21b60ae79e.2sba6cpuq76jpx3j.rcgu.o):
>   in function `nitrokey::get_library_version':
>     nitrokey-rs/src/lib.rs:623: undefined reference to `NK_get_library_version'

It's not exactly clear why that is: All generated rlibs seem to have the
proper symbols in both cases, with and without LTO. That leaves tooling
incompatibilities as a potential culprit [1], but superficially that did
not seem to be the case either, as swapping out ar for gcc-ar did not
make any difference.
To fix this problem, this change proposes the addition of the -fno-lto
flag to the libnitrokey build. The size of the nitrocli binary built in
release mode, i.e., with full LTO enabled through Cargo but without any
CXXFLAG dance, yields exactly the same size with and without this
change, suggesting that this addition does not interfere with the magic
that Cargo does when doing LTO enabled builds.

[0]: https://bugs.gentoo.org/854228
[1]: https://stackoverflow.com/questions/25878407/how-can-i-use-lto-with-static-libraries